Fix tests to run correctly using MacOS `wc` (#106)
As noticed in #105, MacOS `wc -l` produces extra whitespace padding at
the front of the word count.  This makes it so that running the tests on
MacOS produces false negatives.

The fix here is to add a new function that calls `wc -l` then uses `sed`
to strip out the extra whitespace.  This is only used in the tests.
After running this on a MacOS machine, all tests pass as expected.